CONSUPAI-206: agents:fix: rename activitiy_check to activity_check for consistency in imports
